Restrict access-level of variables and methods (#55)
"lightstep.*" keys should for the most part be handled internally
by lightstep and not set by clients. Thus these variables are now
made private or protected. The exceptions being compoent_name
and guid, which can be set by clients via Options.

In some cases methods appeared not to be in use, but are part of
the API, these have been marked with @SuppressWarnings("unused")
to make it clear they are part of the API.