Update compileSdkVersion and buildTools to latest.
It's recommended for Android developers that they use the latest compileSdkVersion
and build tools at all times. This provides newer and better compiler features.

Lint still warns based on minimumSdkVersion, so we needn't worry about using features
that are incompatible with other devices even if we update compileSdkVersion and
buildTools.