Fix dependency scope of maven local publish for jre and common
./gradlew publishToMavenLocal creates poms for both java-common and
lightstep-tracer-jre that reference all dependencies as 'runtime'.
When publishing however we publish them all as 'compile', which
is the desired scope.

I found this article about how to fix it:
https://discuss.gradle.org/t/maven-publish-plugin-generated-pom-making-dependency-scope-runtime/7494

Although very old, it still seems relevant and works for us.

This change has no effect on the published artifacts in bintray and Maven
central, only on the artifacts if published locally.